feat: move jwt endpoints to separate folder, and create additional endpoints
