+++ db/session.py
@@ -10,7 +10,6 @@ import settings
 # BLOCK FOR COMMON INTERACTION WITH DATABASE #
 ##############################################
 
-# create async engine for interaction with database
 engine = create_async_engine(
     settings.REAL_DATABASE_URL,
     future=True,
@@ -18,12 +17,10 @@ engine = create_async_engine(
     execution_options={"isolation_level": "AUTOCOMMIT"},
 )
 
-# create session for the interaction with database
 async_session = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
 
 
 async def get_db() -> Generator:
-    """Dependency for getting async session"""
     try:
         session: AsyncSession = async_session()
         yield session
